+++ src/lib.rs
@@ -87,9 +87,9 @@
 //! components.
 //!
 
-pub type Vec3 = geo_nd::FArray<f32, 3>;
-pub type Vec4 = geo_nd::FArray<f32, 4>;
-pub type Quat = geo_nd::QArray<f32, Vec3, Vec4>;
+pub type Vec3 = geo_nd::FArray<f64, 3>;
+pub type Vec4 = geo_nd::FArray<f64, 4>;
+pub type Quat = geo_nd::QArray<f64, Vec3, Vec4>;
 
 mod subcube;
 pub use subcube::Subcube;
