+++ tests/iau.rs
@@ -10,8 +10,8 @@ fn test_iau() -> Result<(), Box<dyn Error>> {
     catalog.derive_data();
     eprintln!("Loaded {} stars", catalog.len());
     for (name, opt_id, ra, de) in iau::NAMES_AND_RA_DE.iter() {
-        let ra = (*ra as f32) / 180.0 * std::f32::consts::PI;
-        let de = (*de as f32) / 180.0 * std::f32::consts::PI;
+        let ra = (*ra as f64) / 180.0 * std::f64::consts::PI;
+        let de = (*de as f64) / 180.0 * std::f64::consts::PI;
         let (c, star) = catalog.closest_to(ra, de).unwrap();
         let found_id = catalog.star(star).id();
         if let Some(iau_id) = *opt_id {
